Refactored Views into reusable layouts and components
