+++ src/views/components/Calculator.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Calculator extends React.Component {
+    render() {
+        return (
+            <div>
+                <h2>Main Embedded Calculator here</h2>
+            </div>
+        );
+    }
+}
+
+export default Calculator;
