+++ src/views/layouts/Footer.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+class Footer extends React.Component {
+  render() {
+    return (
+      <footer>
+        <p>Footer goes here</p>
+      </footer>
+    );
+  }
+}
+
+export default Footer;
